docs: Escape double curlies in 4.16.0 release notes
This is a one-off workaround. If future commit messages contain double
curlies, we'd have to do this again.

I looked at disabling templating in the blog post Markdown files but
couldn't figure out how to do it while preserving templating for the
component library Markdown files. If we were willing to write the
component library in HTML, we could disable Markdown templating.