chore: Enable Nunjucks throwOnUndefined (#253)
* fix: Incorrect translation reference

`categories` is a sibling rather than a child of `search`.

* fix: Typoed teaser key in 1.0.0-rc-1 frontmatter

* docs: Restore original 1.0.0-rc-1 teaser

While I was fixing the `frontmatter.teaser` typo, I noticed that this
post's teaser belonged to the 1.0.0-rc-2 post. I restored the post's
original teaser, which I found at
https://github.com/eslint/archive-website/blame/bc1052d7b9c5b671c99fd1f8475bb1534937ba34/_posts/2015-07-15-eslint-1.0.0-rc-1-released.md#L15.

* docs: Escape double curlies in 4.16.0 release notes

This is a one-off workaround. If future commit messages contain double
curlies, we'd have to do this again.

I looked at disabling templating in the blog post Markdown files but
couldn't figure out how to do it while preserving templating for the
component library Markdown files. If we were willing to write the
component library in HTML, we could disable Markdown templating.

* fix: Pass sponsors page title from translations

Currently, https://eslint.org/sponsors/ is missing the hero area title.

* docs: Add fallback anonymous sponsor name

* fix: Sort library components by title

Without this sort, each time we edit a component's file, Eleventy will
by default sort it to the bottom of the navigation list. I discovered
this after I edited the Card docs.

* fix: Add missing url to card macro examples

`url` is a required parameter for this macro.

* docs: Add title to Component Library index page

The components layout assumes a title exists, and this will make the
index page consistent with the components themselves.

* chore: Enable Nunjucks throwOnUndefined

As you've seen in the previous commits, this will throw if we're missing
any parameters or accidentally reference substitutions that don't exist.

* docs: Add French download_svg translation

* fix: Typoed footer.social_icons.mailing_list pt-br translation